Conflicts:
	src/main/resources/database/mysql/alter_15_pre_0_to_15_pre_4.sql
	src/main/resources/database/oracle/alter_15_pre_0_to_15_pre_4.sql
	src/main/resources/database/postgresql/alter_15_pre_0_to_15_pre_4.sql
	src/main/webapp/static/themes/light/theme.css
	src/main/webapp/static/themes/light/theme.css.map
	src/main/webapp/static/themes/openolat/theme.css
	src/main/webapp/static/themes/openolat/theme.css.map

+++ src/main/java/org/olat/group/model/StatisticsBusinessGroupRow.java
@@ -41,19 +41,19 @@ public class StatisticsBusinessGroupRow extends BusinessGroupRow {
 		numPending = pending == null ? 0 : pending.intValue();
 	}
 
-	public long getNumOfCoaches() {
+	public int getNumOfCoaches() {
 		return numOfCoaches;
 	}
 	
-	public long getNumOfParticipants() {
+	public int getNumOfParticipants() {
 		return numOfParticipants;
 	}
 	
-	public long getNumWaiting() {
+	public int getNumWaiting() {
 		return numWaiting;
 	}
 
-	public long getNumPending() {
+	public int getNumPending() {
 		return numPending;
 	}
